Update libsais to v2.7.3
Update the `libsais` submodule to current master/HEAD, which is v2.7.3
plus a small fix to `CMakeLists.txt` (79cb777) and a reference to the
`libcubwt` project in `README.md` (d5e74eb).

Also, change the version of the `sais-sys` crate to reflect the libsais
version.

Signed-off-by: Will Woods <w@wizard.zone>

fixup: also change crate version to match libsais version